[configurator] Fix frame timing issue
Fixes #18.

+++ configurator/ee/lib/imgui/backends/imgui_impl_ps2sdk.cpp
@@ -26,7 +26,7 @@
 struct ImGui_ImplPs2Sdk_Data
 {
     GSGLOBAL* Global;
-    u64 Time;
+    float Time;
 
     ImGui_ImplPs2Sdk_Data()   { memset(this, 0, sizeof(*this)); }
 };
